Capture unmatched enums in wrapping data structure
